fix: remove DRY_RUN env var and --dry-run flag (#210)
Free the DRY_RUN env var name to be used for other purposes by function
authors.

The original intent of the DRY_RUN was to be used at build time for GCF
to validate function syntax without starting the server, but this was
never implemented. For local testing purposes, simply starting the
functions framework server is a better method.