fix(ci): Add release-assets.githubusercontent.com to allowed endpoints (#394)
The CodeQL workflow was failing because the `step-security/harden-runner` was blocking egress traffic to the endpoint used for downloading the CodeQL bundle.

The download from `https://github.com/github/codeql-action/releases` redirects to `release-assets.githubusercontent.com`.

This change adds `release-assets.githubusercontent.com` to the list of allowed endpoints to resolve the `ECONNREFUSED` error and removes previous incorrect attempts.